Drop requirement for instance_store_volumes in order to proceed
I can't see why this is still needed. It forces us to check a hardcoded list of instance types
in the Fog Gem that is sometimes out of date (like right now) and it prevents us from handling
up EBS/NVME-storage instances. There may be a reason this bites us but I can't figure out what
it is.